[BDAP] Move directory code to it's own directory
