+++ src/bdap/directory.h
